refactor : 커넥터 패키지 위치 이동
