+++ src/main/java/persistence/connector/DataBaseConnector.java
@@ -1,4 +1,4 @@
-package persistence.transaction;
+package persistence.connector;
 
 import java.sql.Connection;
