+++ src/main/java/persistence/connector/MySqlConnector.java
@@ -1,4 +1,4 @@
-package persistence.transaction;
+package persistence.connector;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
